final commit with thhis version. Although it is able to handle multiple file descriptors, when BUFFER_SIZE is very large then memory allocation fails, not to mention previously mentioned inefficient 1 byte reading.

+++ get_next_line/version0/get_next_line_utils.c
@@ -4,8 +4,8 @@
 char	*new_str(char *buffer, char *str, int i, int reset_byte_count)
 {
 	char		*new_str;
-	int			j;
-	int			k;
+	size_t			j;
+	size_t			k;
 	static int	byte_count = 0;
 
 	if (reset_byte_count > 0)
